refactor: move GetAuthorizedResources into tokens.go
