+++ routers/api/v2/tokens.go
@@ -2,11 +2,14 @@ package v2
 
 import (
 	"encoding/json"
+	"fmt"
 	"github.com/gin-gonic/gin"
+	"github.com/pkg/errors"
 	v2 "github.com/unicsmcr/hs_auth/authorization/v2"
 	"github.com/unicsmcr/hs_auth/routers/api/models"
 	"go.uber.org/zap"
 	"net/http"
+	"net/url"
 	"strings"
 )
 
@@ -62,5 +65,57 @@ func (r *apiV2Router) CreateServiceToken(ctx *gin.Context) {
 
 	ctx.JSON(http.StatusOK, serviceTokenRes{
 		Token: token,
+	})
+}
+
+// GET: /api/v2/tokens/resources/authorized?from={authorisedUris}
+// Request:	 authorisedUris string
+// Response: authorisedUris []v2.UniformResourceIdentifier
+// Headers:  Authorization <- token
+func (r *apiV2Router) GetAuthorizedResources(ctx *gin.Context) {
+	fromUris := ctx.Query("from")
+	if len(fromUris) == 0 {
+		r.logger.Debug("could not parse get authorized resources request")
+		models.SendAPIError(ctx, http.StatusBadRequest, "failed to parse request")
+		return
+	}
+
+	fromUris, err := url.QueryUnescape(fromUris)
+	if err != nil {
+		r.logger.Debug("could not parse get authorized resources request", zap.Error(err))
+		models.SendAPIError(ctx, http.StatusBadRequest, "failed to parse request")
+		return
+	}
+
+	// Remove the leading '[' and trailing ']' from the uri parameter
+	fromUris = fromUris[1 : len(fromUris)-1]
+	rawUriList := strings.Split(fromUris, ",")
+
+	var requestedUris = make([]v2.UniformResourceIdentifier, len(rawUriList))
+	for i, rawUri := range rawUriList {
+		err := requestedUris[i].UnmarshalJSON([]byte(rawUri))
+		if err != nil {
+			r.logger.Debug(fmt.Sprintf("uri could not be parsed at index %d", i))
+			models.SendAPIError(ctx, http.StatusBadRequest, "provided uri could not be parsed")
+			return
+		}
+	}
+
+	token := r.GetAuthToken(ctx)
+	authorizedUris, err := r.authorizer.GetAuthorizedResources(token, requestedUris)
+	if err != nil {
+		switch errors.Cause(err) {
+		case v2.ErrInvalidToken:
+			r.logger.Debug("invalid token", zap.Error(err))
+			r.HandleUnauthorized(ctx)
+		default:
+			r.logger.Error("could not get authorized URIs", zap.Error(err))
+			models.SendAPIError(ctx, http.StatusInternalServerError, "something went wrong")
+		}
+		return
+	}
+
+	ctx.JSON(http.StatusOK, getAuthorizedResourcesRes{
+		AuthorizedUris: authorizedUris,
 	})
 }
